TPS_Predictor_v1_Original/requirements.txt
biopython==1.81
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.65.0
joblib==1.3.2
xgboost==1.7.6
lightgbm==4.0.0
torch==2.0.1
transformers==4.33.2
